# openapi.yaml
openapi: 3.1.0
info:
  title: Cleaning Service API
  description: Backend API for cleaning service management with order processing, pricing calculations, and Google Calendar integration
  version: 1.0.0
  contact:
    name: API Support
    email: support@cleaningservice.com
servers:
  - url: https://api.cleaningservice.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create new order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '422':
          description: Invalid timeslot or service parameters
  /api/orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order details
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Unauthorized
        '404':
          description: Order not found
  /api/orders/slots:
    get:
      tags: [Orders]
      summary: Get available timeslots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-15"
      responses:
        '200':
          description: Available timeslots retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeslotsResponse'
        '400':
          description: Invalid date format
  /api/orders/calc:
    post:
      tags: [Orders]
      summary: Calculate order price and duration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCalculation'
      responses:
        '200':
          description: Calculation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResponse'
        '400':
          description: Validation error

  # Admin endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ClientResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        full_name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone:
          type: string
          example: "+1234567890"
        address:
          type: string
          example: "123 Main St, City, State 12345"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
...
    ServiceResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Couch Cleaning"
        description:
          type: string
          example: "Professional deep cleaning for all types of couches"
        category:
          type: string
          enum: [couch, rug, window, other]
          example: "couch"
        price_per_removable_cushion:
          type: number
          minimum: 0
          example: 30.0
        price_per_unremovable_cushion:
          type: number
          minimum: 0
          example: 18.0
        price_per_pillow:
          type: number
          minimum: 0
          example: 5.0
        price_per_window:
          type: number
          minimum: 0
          example: 25.0
        base_surcharge_pct:
          type: number
          minimum: 0
          example: 38.0
        pet_hair_surcharge_pct:
          type: number
          minimum: 0
          example: 15.0
        urine_stain_surcharge_pct:
          type: number
          minimum: 0
          example: 5.0
        accelerated_drying_surcharge:
          type: number
          minimum: 0
          example: 45.0
        before_image:
          type: string
          format: uri
          nullable: true
        after_image:
          type: string
          format: uri
          nullable: true
        is_published:
          type: boolean
          example: true
...
    CleanerResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        full_name:
          type: string
          example: "Jane Smith"
        phone:
          type: string
          example: "+1234567890"
        email:
          type: string
          format: email
          example: "jane.smith@cleaningservice.com"
        services:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
        calendar_email:
          type: string
          format: email
          nullable: true
          example: "jane.smith@cleaningservice.com"
...
    OrderStatus:
      type: string
      enum: [Pending Confirmation, Confirmed, Completed, Cancelled]
      example: "Pending Confirmation"
    OrderCreate:
      type: object
      required:
        - scheduled_date
        - scheduled_time
        - order_items
        - notes
      properties:
        scheduled_date:
          type: string
          format: date
          example: "2024-01-20"
        scheduled_time:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "14:00"
        order_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        notes:
          type: string
          maxLength: 1000
          example: "Please use eco-friendly cleaning products"
    OrderItemCreate:
      type: object
      required:
        - service_id
        - parameters
      properties:
        service_id:
          type: integer
          example: 1
        parameters:
          $ref: '#/components/schemas/ServiceParameters'
    ServiceParameters:
      type: object
      properties:
        removable_cushion_count:
          type: integer
          minimum: 0
          example: 2
        unremovable_cushion_count:
          type: integer
          minimum: 0
          example: 1
        pillow_count:
          type: integer
          minimum: 0
          example: 4
        window_count:
          type: integer
          minimum: 0
          example: 3
        rug_width:
          type: number
          minimum: 0
          example: 8.0
        rug_length:
          type: number
          minimum: 0
          example: 10.0
        rug_count:
          type: integer
          minimum: 1
          example: 1
        base_cleaning:
          type: boolean
          example: true
        pet_hair:
          type: boolean
          example: false
        urine_stains:
          type: boolean
          example: false
        accelerated_drying:
          type: boolean
          example: false
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        client:
          $ref: '#/components/schemas/ClientResponse'
        scheduled_date:
          type: string
          format: date
          example: "2024-01-20"
        scheduled_time:
          type: string
          example: "14:00"
        total_duration_minutes:
          type: integer
          example: 180
        total_price:
          type: number
          example: 245.50
        status:
          $ref: '#/components/schemas/OrderStatus'
        cleaner:
          $ref: '#/components/schemas/CleanerResponse'
          nullable: true
        notes:
          type: string
          example: "Please use eco-friendly cleaning products"
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        service:
          $ref: '#/components/schemas/ServiceResponse'
        parameters:
          $ref: '#/components/schemas/ServiceParameters'
        calculated_cost:
          type: number
          example: 120.75
        calculated_time_minutes:
          type: integer
          example: 90
...
    OrderCalculation:
      type: object
      required:
        - order_items
      properties:
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
    CalculationResponse:
      type: object
      properties:
        total_price:
          type: number
          example: 245.50
        total_duration_minutes:
          type: integer
          example: 180
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'

    # Timeslot schemas
    TimeslotsResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2024-01-15"
        available_slots:
          type: array
          items:
            type: string
            pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: ["10:00", "10:30", "11:00", "14:00", "14:30"]
        working_hours:
          type: object
          properties:
            start:
              type: string
              example: "10:00"
            end:
              type: string
              example: "19:00"
        slot_duration_minutes:
          type: integer
          example: 30

    # Error schemas
...